feat: track approval methods
