+++ src/modules/translation/data-access/GlossRepository.ts
@@ -1,18 +1,5 @@
 import { query } from "@/db";
-
-export enum GlossStateRaw {
-  Approved = "APPROVED",
-  Unapproved = "UNAPPROVED",
-}
-export enum GlossSourceRaw {
-  User = "USER",
-  Import = "IMPORT",
-}
-export enum GlossApprovalMethodRaw {
-  UserInput = "USER_INPUT",
-  GoogleSuggestion = "GOOGLE_SUGGESTION",
-  MachineSuggestion = "MACHINE_SUGGESTION",
-}
+import { GlossSourceRaw, GlossStateRaw } from "../types";
 
 export interface DbGloss {
   gloss: string | null;
