+++ src/modules/translation/types.ts
@@ -1,0 +1,13 @@
+export enum GlossStateRaw {
+  Approved = "APPROVED",
+  Unapproved = "UNAPPROVED",
+}
+export enum GlossSourceRaw {
+  User = "USER",
+  Import = "IMPORT",
+}
+export enum GlossApprovalMethodRaw {
+  UserInput = "USER_INPUT",
+  GoogleSuggestion = "GOOGLE_SUGGESTION",
+  MachineSuggestion = "MACHINE_SUGGESTION",
+}
